Abgabe UE3: Refactor currency conversion and update tests
The currency conversion methods in the Girokonto and Sparbuch classes have been refactored to remove redundant parameters. Currency rounding and accuracy have been improved in the Waehrung class. A new KontoTest file was added with extensive testing for currency conversion. The version in pom.xml was updated, and various dependencies have been added for testing and better calculations.