Correction feedback implemented
Account number is no longer generated with a random number, but is generated by an iteration of the account number error. In addition, the recursive call has been removed and replaced by a while loop.

In addition, the "validiereKontoUndBetrag" method has been split into "validiereKonto" and "validiereBetrag" to make the check of the two parameters more independent.