Update purchase and sale prices in AktienSpielereien and make executorService non-transient in Konto
Updated the purchase and sale prices for Apple and Microsoft stocks in the AktienSpielereien.java. Also ensured that the executorService in Konto.java is non-transient. This allows the executorService to be serialized along with other fields of a Konto object. A readObject() method was added to Konto.java to ensure that the executorService is properly initialized during deserialization.