Refactor deposit and withdraw methods in KontoController
The deposit and withdrawal methods in the KontoController class have been updated to take a double value as a parameter instead of retrieving it internally from the view to improve the separation of concerns. Correspondingly, update handlers in the KontoOberflaeche class to parse the text field value and pass it to the controller methods, adding error handling for non-parsable values. Also removed getBetrag method from the KontoOberflaeche class.